+++ src/stores/userinfo.js
@@ -1,0 +1,13 @@
+import { defineStore } from 'pinia'
+
+export const userInfoStore = defineStore('userInfo', {
+    state: () => ({
+        userInfo: {
+            username: 'John Doe',
+            telegram:  undefined,
+        },
+    }),
+    getters: {
+        getUserInfo: (state) => state.userInfo,
+    }
+})
