+++ src/utils/request.js
@@ -1,10 +1,12 @@
 // src/utils/request.js
 import axios from 'axios';
 
+
+
 // 创建实例
 const service = axios.create({
-    baseURL: import.meta.env.VITE_API_BASE_URL, // Vite 项目用 env
-    timeout: 10000, // 超时 10 秒
+    baseURL: '/api', // ✅ 使用代理路径
+    timeout: 30000,
 });
 
 // 请求拦截器
@@ -19,8 +21,8 @@ service.interceptors.request.use(
         return config;
     },
     (error) => {
-        console.error('请求错误:', error);
-        return Promise.reject(error);
+
+        return {...error,status:'error',};
     }
 );
 
@@ -31,9 +33,9 @@ service.interceptors.response.use(
         if (res.code !== 0) {
             // 根据后端返回的 code 判断成功/失败
             console.warn('接口报错:', res.message || '接口错误');
-            return Promise.reject(res);
+            return {...res,status:'error'};
         }
-        return res; // 返回 data 部分
+        return res.data; // 返回 data 部分
     },
     (error) => {
         console.error('响应错误:', error);
@@ -46,7 +48,7 @@ service.interceptors.response.use(
                 alert('服务器错误，请稍后再试');
             }
         }
-        return Promise.reject(error);
+        return {...error,status:'errors'};
     }
 );
